requirements.txt
beautifulsoup4==4.11.1
fitz==0.0.1.dev2
pdf2image==1.16.0
pdfkit==1.0.0
Pillow==9.1.0
PyMuPDF==1.19.4
requests==2.22.0
